Rename HTMLElement as DOMElement plugin
